+++ finagle-base-http/src/main/scala/com/twitter/finagle/http/headers/DefaultHeaderMap.scala
@@ -1,4 +1,4 @@
-package com.twitter.finagle.http
+package com.twitter.finagle.http.headers
 
 object DefaultHeaderMap {
   def apply(headers: (String, String)*): HeaderMap = HashBackedHeaderMap(headers:_*)
