+++ finagle-base-http/src/main/scala/com/twitter/finagle/http/headers/Rfc7230HeaderValidation.scala
@@ -1,4 +1,4 @@
-package com.twitter.finagle.http
+package com.twitter.finagle.http.headers
 
 import java.util.BitSet
