+++ finagle-base-http/src/test/scala/com/twitter/finagle/http/headers/DefaultHeaderMapTest.scala
@@ -1,4 +1,4 @@
-package com.twitter.finagle.http
+package com.twitter.finagle.http.headers
 
 class DefaultHeaderMapTest extends AbstractHeaderMapTest {
   final def newHeaderMap(headers: (String, String)*): HeaderMap = DefaultHeaderMap(headers: _*)
