+++ finagle-base-http/src/test/scala/com/twitter/finagle/http/headers/Rfc7230HeaderValidationTest.scala
@@ -1,6 +1,6 @@
-package com.twitter.finagle.http
+package com.twitter.finagle.http.headers
 
-import com.twitter.finagle.http.Rfc7230HeaderValidation.{ObsFoldDetected, ValidationFailure}
+import com.twitter.finagle.http.headers.Rfc7230HeaderValidation.{ObsFoldDetected, ValidationFailure}
 import org.scalacheck.Gen
 import org.scalatest.FunSuite
 import org.scalatestplus.scalacheck.ScalaCheckDrivenPropertyChecks
